Version 4.1
Trying to make it work with Observable and Observer so that you can
pick up and drop items. I think I got halfway but do not have enough
time today to finish it. Feel free to keep on trying with my my version
or use Version 4 to start fresh and implement your own Observers!